+++ src/Open3D/Types/Vector3f.h
@@ -8,17 +8,17 @@ typedef struct Vector3f {
     static const uint ROWS = 1;
     static const uint COLS = 3;
 
-    double s[COLS];
+    float s[COLS];
 
     // subscript operator: readwrite
-    double &operator[](const uint &i);
+    float &operator[](const uint &i);
     // subscript operator: readonly
-    const double &operator[](const uint &i) const;
+    const float &operator[](const uint &i) const;
 
     // casting operator: readwrite
-    explicit operator double *const();
+    explicit operator float *const();
     // casting operator: readonly
-    explicit operator const double *const();
+    explicit operator const float *const();
 } Vector3f;
 
     bool operator==(const Vector3f &v0, const Vector3f &v1);
@@ -35,19 +35,19 @@ typedef struct Vector3f {
     // subtraction assignment
     Vector3f &operator-=(Vector3f &v0, const Vector3f &v1);
     // addition
-    Vector3f operator+(const Vector3f &v, const double &t);
+    Vector3f operator+(const Vector3f &v, const float &t);
     // subtraction
-    Vector3f operator-(const Vector3f &v, const double &t);
+    Vector3f operator-(const Vector3f &v, const float &t);
     // multiply with scalar
-    Vector3f operator*(const Vector3f &v, const double &t);
+    Vector3f operator*(const Vector3f &v, const float &t);
     // divide by scalar
-    Vector3f operator/(const Vector3f &v, const double &t);
+    Vector3f operator/(const Vector3f &v, const float &t);
     // addition assignment
-    Vector3f &operator+=(Vector3f &v, const double &t);
+    Vector3f &operator+=(Vector3f &v, const float &t);
     // subtraction assignment
-    Vector3f &operator-=(Vector3f &v, const double &t);
+    Vector3f &operator-=(Vector3f &v, const float &t);
     // multiplication assignment
-    Vector3f &operator*=(Vector3f &v, const double &t);
+    Vector3f &operator*=(Vector3f &v, const float &t);
     // division assignment
-    Vector3f &operator/=(Vector3f &v, const double &t);
+    Vector3f &operator/=(Vector3f &v, const float &t);
 }  // namespace open3d
